# core/state: optimize some internals during encoding