--- a/core/state/state_object_test.go
+++ b/core/state/state_object_test.go
@@ -1,0 +1,70 @@
+// Copyright 2019 The go-ethereum Authors
+// This file is part of the go-ethereum library.
+//
+// The go-ethereum library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The go-ethereum library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
+
+package state
+
+import (
+	"bytes"
+	"fmt"
+	"math/rand"
+	"testing"
+	"time"
+
+	"github.com/ethereum/go-ethereum/common"
+)
+
+func BenchmarkCutOriginal(b *testing.B) {
+	value := common.HexToHash("0x01")
+	for i := 0; i < b.N; i++ {
+		bytes.TrimLeft(value[:], "\x00")
+	}
+}
+
+func BenchmarkCutsetterFn(b *testing.B) {
+	value := common.HexToHash("0x01")
+	cutSetFn := func(r rune) bool {
+		return int32(r) == int32(0)
+	}
+	for i := 0; i < b.N; i++ {
+		bytes.TrimLeftFunc(value[:], cutSetFn)
+	}
+}
+
+func BenchmarkCutCustomTrim(b *testing.B) {
+	value := common.HexToHash("0x01")
+	for i := 0; i < b.N; i++ {
+		common.TrimLeftZeroes(value[:])
+	}
+}
+
+func xTestFuzzCutter(t *testing.T) {
+	rand.Seed(time.Now().Unix())
+	for {
+		v := make([]byte, 20)
+		zeroes := rand.Intn(21)
+		rand.Read(v[zeroes:])
+		exp := bytes.TrimLeft(v[:], "\x00")
+		got := common.TrimLeftZeroes(v)
+		if !bytes.Equal(exp, got) {
+
+			fmt.Printf("Input %x\n", v)
+			fmt.Printf("Exp %x\n", exp)
+			fmt.Printf("Got %x\n", got)
+			t.Fatalf("Error")
+		}
+		//break
+	}
+}