I noticed this wile investigating the state a bit. What we currently do before RLP-encoding a slot value is to call `TrimLeft`:
```golang
func TrimLeft(s []byte, cutset string) []byte {
	return TrimLeftFunc(s, makeCutsetFunc(cutset))
}
```
... which dynamically creates a `cutSetFunction` . 
Then it calls `TrimLeftFunc`, which: 
```golang
// TrimLeftFunc treats s as UTF-8-encoded bytes and returns a subslice of s by slicing off
// all leading UTF-8-encoded code points c that satisfy f(c).
func TrimLeftFunc(s []byte, f func(r rune) bool) []byte {
```
It interprets the data as `UTF-8` encoded data, and parses runes. 

In our case, it's the wrong abstraction, what we want to do here is simply trim leading zeroes, which has nothing whatsoever to do with `UTF-8` and runes. 

This PR replaces that with a simple for-loop, more or less. 

benchmark results:
```
[user@work state]$ go test . -bench BenchmarkCut*
OK: 5 passed
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/state
BenchmarkCutOriginal-6     	10000000	       146 ns/op
BenchmarkCutsetterFn-6     	20000000	        86.3 ns/op
BenchmarkCutCustomTrim-6   	50000000	        28.9 ns/op
PASS
ok  	github.com/ethereum/go-ethereum/core/state	11.941s
```
It's probably a drop in the ocean though...