@karalabe 2019-09-10
Linter fails, pls gofmt